requirements.txt
tensorflow-gpu==2.3.0
django==3.2.3
djangorestframework==3.12.4
Pillow==8.2.0
matplotlib==3.4.2